// apps/kanban-web/src/lib/generated/models/DeleteProjectOutputBody.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type DeleteProjectOutputBody = {
    /**
     * A URL to the JSON Schema for this object.
     */
    readonly $schema?: string;
    deleted: boolean;
    project: string;
};